修改数据类型

+++ CyPal/src/main/java/com/cypal/ming/cypal/bean/IndexEntity.java
@@ -162,8 +162,8 @@ public class IndexEntity {
              */
 
             public int todaySuccess;
-            public int todaySuccessMoney;
-            public int todayCommision;
+            public double todaySuccessMoney;
+            public double todayCommision;
         }
 
         public class VersionBean {
